Added downloadable excel file for batch cost tracking

+++ CWBDataApp/models.py
@@ -3,27 +3,27 @@ from django.db import models
 class Batchcosttracking(models.Model):
     batchname = models.CharField(db_column='BatchName', primary_key=True, max_length=15)
     batchdate = models.DateField(db_column='BatchDate', blank=True, null=True)
-    totalcost = models.FloatField(db_column='TotalCost')
+    totalcost = models.DecimalField(db_column='TotalCost', decimal_places=2, max_digits=10)
     totalweight = models.SmallIntegerField(db_column='TotalWeight')
-    priceperpound = models.FloatField(db_column='PricePerPound')
+    priceperpound = models.DecimalField(db_column='PricePerPound', decimal_places=2, max_digits=10)
     supplier1 = models.ForeignKey('Materialcost', models.DO_NOTHING, db_column='Supplier1',  related_name='supplier1')
     weight1 = models.SmallIntegerField(db_column='Weight1', blank=True, null=True)
-    value1 = models.FloatField(db_column='Value1', blank=True, null=True)
+    value1 = models.DecimalField(db_column='Value1', blank=True, null=True, decimal_places=3, max_digits=10)
     supplier2 = models.ForeignKey('Materialcost', models.DO_NOTHING, db_column='Supplier2', related_name='supplier2', blank=True, null=True)
     weight2 = models.SmallIntegerField(db_column='Weight2', blank=True, null=True)
-    value2 = models.FloatField(db_column='Value2', blank=True, null=True)
+    value2 = models.DecimalField(db_column='Value2', blank=True, null=True, decimal_places=3, max_digits=10)
     supplier3 = models.ForeignKey('Materialcost', models.DO_NOTHING, db_column='Supplier3', related_name='supplier3', blank=True, null=True)
     weight3 = models.SmallIntegerField(db_column='Weight3', blank=True, null=True)
-    value3 = models.FloatField(db_column='Value3', blank=True, null=True)
+    value3 = models.DecimalField(db_column='Value3', blank=True, null=True,  decimal_places=3, max_digits=10)
     supplier4 = models.ForeignKey('Materialcost', models.DO_NOTHING, db_column='Supplier4', related_name='supplier4', blank=True, null=True)
     weight4 = models.SmallIntegerField(db_column='Weight4', blank=True, null=True)
-    value4 = models.FloatField(db_column='Value4', blank=True, null=True)
+    value4 = models.DecimalField(db_column='Value4', blank=True, null=True,  decimal_places=3, max_digits=10)
     colour = models.CharField(db_column='Colour', max_length=20, blank=True, null=True)
-    colourweight = models.FloatField(db_column='ColourWeight', blank=True, null=True)
-    colourvalue = models.FloatField(db_column='ColourValue', blank=True, null=True)
+    colourweight = models.DecimalField(db_column='ColourWeight', blank=True, null=True, decimal_places=2, max_digits=10)
+    colourvalue = models.DecimalField(db_column='ColourValue', blank=True, null=True, decimal_places=3, max_digits=10)
     foam = models.CharField(db_column='Foam', max_length=4, blank=True, null=True)
-    foamweight = models.FloatField(db_column='FoamWeight', blank=True, null=True)
-    foamvalue = models.FloatField(db_column='FoamValue', blank=True, null=True)
+    foamweight = models.DecimalField(db_column='FoamWeight', blank=True, null=True, decimal_places=2, max_digits=10)
+    foamvalue = models.DecimalField(db_column='FoamValue', blank=True, null=True,  decimal_places=3, max_digits=10)
     totalshredweight = models.SmallIntegerField(db_column='TotalShredWeight', blank=True, null=True)
 
     class Meta:
@@ -33,7 +33,7 @@ class Batchcosttracking(models.Model):
 
 class Materialcost(models.Model):
     supplier = models.CharField(db_column='Supplier', primary_key=True, max_length=30)
-    costperpound = models.FloatField(db_column='CostPerPound')
+    costperpound = models.DecimalField(db_column='CostPerPound', decimal_places=3, max_digits=10)
 
     class Meta:
         managed = False
@@ -46,7 +46,7 @@ class Materialinventory(models.Model):
     numberofboxes = models.SmallIntegerField(db_column='NumberofBoxes')
     locations = models.CharField(db_column='Locations', max_length=20, blank=True, null=True)
     premixed = models.BooleanField(db_column='PreMixed')
-    priceperpound = models.FloatField(db_column='PricePerPound')
+    priceperpound = models.DecimalField(db_column='PricePerPound', decimal_places=2, max_digits=10)
 
     class Meta:
         managed = False
@@ -123,7 +123,7 @@ class Ordersheetmachine1(models.Model):
     projectedstartdate = models.DateField(db_column='ProjectedStartDate')
     projectedenddate = models.DateField(db_column='ProjectedEndDate')
     duedate = models.DateField(db_column='DueDate')
-    lengthofrunindays = models.FloatField(db_column='LengthofRuninDays', blank=True, null=True)
+    lengthofrunindays = models.DecimalField(db_column='LengthofRuninDays', blank=True, null=True,  decimal_places=2, max_digits=10)
     priorityrank = models.SmallIntegerField(db_column='PriorityRank')
     boardprofile = models.ForeignKey('Productinventory', models.DO_NOTHING, db_column='BoardProfile', related_name='Machine1boardprofile')
     colour = models.CharField(db_column='Colour', max_length=15)
@@ -145,7 +145,7 @@ class Ordersheetmachine2(models.Model):
     projectedstartdate = models.DateField(db_column='ProjectedStartDate')
     projectedenddate = models.DateField(db_column='ProjectedEndDate')
     duedate = models.DateField(db_column='DueDate')
-    lengthofrunindays = models.FloatField(db_column='LengthofRuninDays', blank=True, null=True)
+    lengthofrunindays = models.DecimalField(db_column='LengthofRuninDays', blank=True, null=True,  decimal_places=2, max_digits=10)
     priorityrank = models.SmallIntegerField(db_column='PriorityRank')
     boardprofile = models.ForeignKey('Productinventory', models.DO_NOTHING, db_column='BoardProfile', related_name='Machine2boardprofile')
     colour = models.CharField(db_column='Colour', max_length=15)
@@ -167,7 +167,7 @@ class Ordersheetmachine3(models.Model):
     projectedstartdate = models.DateField(db_column='ProjectedStartDate')
     projectedenddate = models.DateField(db_column='ProjectedEndDate')
     duedate = models.DateField(db_column='DueDate')
-    lengthofrunindays = models.FloatField(db_column='LengthofRuninDays', blank=True, null=True)
+    lengthofrunindays = models.DecimalField(db_column='LengthofRuninDays', blank=True, null=True, decimal_places=2, max_digits=10)
     priorityrank = models.SmallIntegerField(db_column='PriorityRank')
     boardprofile = models.ForeignKey('Productinventory', models.DO_NOTHING, db_column='BoardProfile', related_name='Machine3boardprofile')
     colour = models.CharField(db_column='Colour', max_length=15)
